style: preliminary navbar style
