+++ src/App.css
@@ -26,6 +26,3 @@
   transform: translate(-50%, -50%);
 }
 
-.navigation-bar {
-  text-align: center;
-}
