+++ src/NavBar/styles.css
@@ -1,0 +1,32 @@
+.nav-bar {
+  text-align: center;
+  background-image: linear-gradient(90deg, rgba(159,63,234,1) 35%, rgba(176,96,191,1) 84%);
+  height: 40px;
+  display: flex;
+}
+
+.nav-btn {
+    background-color: transparent;
+    border: 1px solid #FFFFFF;
+    transition: background-color 0.5s;
+    border-radius: 10px;
+    font-size: 20px;
+    font-weight: 300;
+    line-height: 30px;
+    color: white;
+}
+
+.nav-btn:hover {
+    background-color: #FFFFFF;
+    color: #9F3FEA;
+}
+
+.nav-btnbar {
+    float: right;
+}
+
+.nav-logo {
+    align-self: left;
+    height: 50px;
+    width: 50px;
+}
